back-end/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
django-filter==24.2